Automatically resolve cache upon evaluation
If a schema with unresolved references is evaluated, first resolve
all references in the schema's cache.  This ensures that reference
resolution problems are detected prior to evaluation proper, at
the cost of an additional failure mode and a potentially
significant performance cost on first evaluation.

Callers can still manually call resolve_references() at either
the schema or catalog level to retain control over this cost and
additional failure mode.  The automatic resolution cost is
significant because resolving only the schemas and references that
are needed creates substantial code complexity.  The difference
is only notable if the schema cache is much larger than the
set of schemas that will be evaluated, which is probably unusual.