Add RewritingSourceMixin & concrete classes
This allows constructing sources that rewrite the URIs requested
so that the URI structure and storage layout can have a more
complex relationship.

This particularly facilitates references to schema resources
embedded in bundles, or references to URI components that use
characters not allowed in filesystems, or a unified URI namespace
that is distributed across multiple repositories in a test environment.