"format" supports all types by default
Previously, formats could set applicable types, but in the decorator
the types were defaulted to ("string",), contrary to the behavior of
keywords in general and the "format" keyword class in particular.

Since the tutorial example, unit tests, and the one format
validator all check the instance type before proceeding, it looked
very much like format validator functions were expected to do
their own type checking.

This change aligns the default type applicability of formats with
that of keywords, and updates all implementations, tests, and
examples to rely on the decorator parameter to handle the type check.