Class attrs for cooperative inheritance
This adds class attributes throughout the package to indicate
which related subclasses should be used.  This is an extension
of earlier work specific to JSONPointer / RelativeJSONPointer.

The number of class attributes per class is kept to a minimum
to ensure that it is as difficult as possible to produce
inconsistent behavior, instead allowing classes to, for example,
find the correct exception class through the most relevant
non-exception class rather than handling them separately.

Setting the `Catalog` subclass is handled slightly differently
due to the need to avoid circular imports, but after initialization
it behaves the same.  Subclasses can either override the class
attribute or the initializer class method.