Add JSON._pre_recurse_init() hook.
Some aspects of document construction require examining parents,
which is problematic when the parent has not had a chance to run
subclass initialization.  But some subclass initialization requires
the basic fields to be set in the superclass.

This adds a hook to run subclass initialization code between the
initial field set-up and the recursive document construction.