Add a pre-recursion constructor hook to JSON
Some aspects of document construction require examining parents,
which is problematic when the parent has not had a chance to run
subclass initialization.  But some subclass initialization requires
the basic fields to be set in the superclass.

This adds a hook to run subclass initialization code between the
initial field and data set-up and the recursive document construction.

The JSON.data attribute is initialized to the passed-in value, prior
to running the hoook, as it is common to need to examine the immediate
scalar fields in a JSON object.  For those fields, the raw value
and the data property behave the same.

While the different behavior of non-scalar fields in this temporary
JSON.data configuration is arguabley problematic, so is the lack
of a data field entirely when trying to do subclass initialization.
All alternate approaches I could come up with either consumed
substantially more memory or became very complex with many points
in the code needing to handle the difference between the
during-__init__() and regular state.  This approach currently works
without causing any new problems.