More flexible (sub)schema/resource instantiation
Allow an arbitrary callable factory to produce new resource
instances in the Catalog; this enables checking the contents
of the document before choosing the class to instantiate.
The ``cls`` parameter is retained to implement the type check,
and as a fallback if no factory is provided.

Add / improve JSONSchema's is_resource_root() and is_format_root().

Allow customizing the subschema/refschema class, which is slightly
more complex than it might first seem because of how the Metaschema
class works as a JSONSchema subclass, and its different
constructor signature.

Both subclass instantiation and instantiation of referenced schemas
uses this customization.

TODO: What if there's a mix of subclasses referencing each other?
It is not clear how / whether that should be supported.